pyproject.toml
[tool.poetry]
name = "full-body-vtuber"
version = "0.1.0"
description = ""
authors = ["megahomyak <g.megahomyak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
mediapipe = "^0.10.8"
opencv-python = "^4.8.1.78"
pyvirtualcam = "^0.10.2"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
